Fix autotitling bug
For some reason a certain ISP has title tags all over their HTML body.
I never thought these were valid semantic tags but eh. Anyway, since our
title grabbing did not restrict the selection to the first found match
it negatively affected user experience. This fix fixes that in two ways,
by restricting it to the first match of a <title> within the <head>
section of the document.

Sure, this did not make it into the final sprint but I just _had_ to fix
it since it bothered me so much at the demo. ":D"